feat: add new lists creation page
